Make toYUYV() and toRGB() return a node Buffer
Previously these functions returned an array of numbers, which took
long time to construct. With this commit, we return a JavaScript
Buffer instead, backed by a C++ SlowBuffer.
On a Raspberry Pi device, before this change toRGB() took ~4200ms,
with this change it's down to 42ms. This relates to issue #9.

This change also adds the function getFrameData(), and toYUYV() is
(incorrectly) an alias to this function.